requirements.txt
# Data
akshare>=1.14.0
baostock>=0.8.0
pandas>=2.0.0
numpy>=1.24.0

# Config
pyyaml>=6.0

# Visualization
streamlit>=1.30.0
plotly>=5.18.0

# Computation
scipy>=1.11.0

# AI Research
openai>=1.0.0

# Logging
loguru>=0.7.0
